Add tests for new container feature
Move container tests to their own file.
