+++ tests/widgets_test.py
@@ -23,7 +23,7 @@ from io import StringIO
 from unittest.mock import patch
 
 from simpleline import App
-from simpleline.render.containers import WindowContainer, ListRowContainer, ListColumnContainer, KeyPattern
+from simpleline.render.containers import ListRowContainer
 from simpleline.render.prompt import Prompt
 from simpleline.render.screen import UIScreen, InputState
 from simpleline.render.widgets import TextWidget, SeparatorWidget, CheckboxWidget, CenterWidget, ColumnWidget
@@ -198,186 +198,6 @@ class Widgets_TestCase(BaseWidgets_TestCase):
         expected_result = [u"   Test"]
 
         self.evaluate_result(w.get_lines(), expected_result)
-
-
-class Containers_TestCase(BaseWidgets_TestCase):
-
-    def _test_callback(self, data):
-        pass
-
-    def test_listrow_container(self):
-        c = ListRowContainer(columns=2, items=[self.w2, self.w3, self.w5], columns_width=10, spacing=2, numbering=False)
-        c.render(25)
-
-        expected_result = [u"Test        Test 2",
-                           u"Test 3"]
-        res_lines = c.get_lines()
-
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_empty(self):
-        c = ListRowContainer(columns=1)
-
-        c.render(10)
-        result = c.get_lines()
-
-        self.assertEqual(len(result), 0)
-
-    def test_more_columns_than_widgets(self):
-        c = ListRowContainer(columns=3, items=[self.w1], columns_width=40, numbering=False)
-        c.render(80)
-
-        expected_result = [u"Můj krásný dlouhý text"]
-
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_listrow_wrapping(self):
-        # spacing is 3 by default
-        c = ListRowContainer(2, [self.w1, self.w2, self.w3, self.w4], columns_width=15, numbering=False)
-        c.render(25)
-
-        expected_result = [u"Můj krásný        Test",
-                           u"dlouhý text",
-                           u"Test 2            Krásný dlouhý",
-                           u"                  text podruhé"]
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_newline_wrapping(self):
-        widgets = [TextWidget("Hello"), TextWidget("Wrap\nthis\ntext"), TextWidget("Hi"),
-                   TextWidget("Hello2")]
-
-        c = ListRowContainer(3, widgets, columns_width=6, spacing=1, numbering=False)
-        c.render(80)
-
-        expected_result = [u"Hello  Wrap   Hi",
-                           u"       this",
-                           u"       text",
-                           u"Hello2"]
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_listcolumn_container(self):
-        c = ListColumnContainer(columns=2, items=[self.w2, self.w3, self.w5], columns_width=10, spacing=2,
-                                numbering=False)
-        c.render(25)
-
-        expected_result = [u"Test        Test 3",
-                           u"Test 2"]
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_listcolumn_wrapping(self):
-        # spacing is 3 by default
-        c = ListColumnContainer(2, [self.w1, self.w2, self.w3, self.w4], columns_width=15, numbering=False)
-        c.render(25)
-
-        expected_result = [u"Můj krásný        Test 2",
-                           u"dlouhý text",
-                           u"Test              Krásný dlouhý",
-                           u"                  text podruhé"]
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_add_new_container(self):
-        c = ListRowContainer(columns=2, items=[TextWidget("Ahoj")], columns_width=15, spacing=0, numbering=False)
-
-        expected_result = [u"Ahoj"]
-
-        c.render(80)
-        self.evaluate_result(c.get_lines(), expected_result)
-
-        c.add(TextWidget("Nový widget"))
-        c.add(TextWidget("Hello"))
-
-        expected_result = [u"Ahoj           Nový widget",
-                           u"Hello"]
-
-        c.render(80)
-        self.evaluate_result(c.get_lines(), expected_result)
-
-    def test_column_numbering(self):
-        # spacing is 3 by default
-        c = ListColumnContainer(2, [self.w1, self.w2, self.w3, self.w4], columns_width=16)
-        c.render(25)
-
-        expected_result = [u"1) Můj krásný      3) Test 2",
-                           u"   dlouhý text",
-                           u"2) Test            4) Krásný dlouhý",
-                           u"                      text podruhé"]
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_row_numbering(self):
-        # spacing is 3 by default
-        c = ListRowContainer(2, [self.w1, self.w2, self.w3, self.w4], columns_width=16)
-        c.render(25)
-
-        expected_result = [u"1) Můj krásný      2) Test",
-                           u"   dlouhý text",
-                           u"3) Test 2          4) Krásný dlouhý",
-                           u"                      text podruhé"]
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_custom_numbering(self):
-        # spacing is 3 by default
-        c = ListRowContainer(2, [self.w1, self.w2, self.w3, self.w4], columns_width=20)
-        c.key_pattern = KeyPattern("a {:d} a ")
-        c.render(25)
-
-        expected_result = [u"a 1 a Můj krásný       a 2 a Test",
-                           u"      dlouhý text",
-                           u"a 3 a Test 2           a 4 a Krásný dlouhý",
-                           u"                             text podruhé"]
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_window_container(self):
-        c = WindowContainer(title="Test")
-
-        c.add(TextWidget("Body"))
-        c.render(10)
-
-        expected_result = [u"Test",
-                           u"",
-                           u"Body"]
-
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_window_container_with_multiple_items(self):
-        c = WindowContainer(title="Test")
-
-        c.add(TextWidget("Body"))
-        c.add(TextWidget("Body second line"))
-        c.render(30)
-
-        expected_result = [u"Test",
-                           u"",
-                           u"Body",
-                           u"Body second line"]
-
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
-
-    def test_window_container_wrapping(self):
-        c = WindowContainer(title="Test")
-
-        c.add(TextWidget("Body long line"))
-        c.add(TextWidget("Body"))
-        c.render(5)
-
-        expected_result = [u"Test",
-                           u"",
-                           u"Body",
-                           u"long",
-                           u"line",
-                           u"Body"]
-
-        res_lines = c.get_lines()
-        self.evaluate_result(res_lines, expected_result)
 
 
 @patch('simpleline.render.io_manager.InOutManager._get_input')
